+++ src/application/components/Queries/__tests__/Queries.test.tsx
@@ -1,26 +1,29 @@
 import React from "react";
-import { screen, within, waitFor } from "@testing-library/react";
-import user from "@testing-library/user-event";
+import { screen, within, waitFor, act } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
 
 import { renderWithApolloClient } from "../../../utilities/testing/renderWithApolloClient";
 import { client, GET_QUERIES } from "../../../index";
 import { Queries } from "../Queries";
-
-jest.mock("../QueryViewer", () => ({
-  QueryViewer: () => null,
-}));
+import type { GetQueries } from "../../../types/gql";
 
 describe("<Queries />", () => {
-  const queries = [
+  const queries: GetQueries["watchedQueries"]["queries"] = [
     {
       id: 0,
       __typename: "WatchedQuery",
       name: null,
+      queryString: "query { hello }",
+      variables: null,
+      cachedData: null,
     },
     {
       id: 1,
       __typename: "WatchedQuery",
       name: "GetColors",
+      queryString: "query GetColors { colors }",
+      variables: null,
+      cachedData: null,
     },
   ];
 
@@ -38,6 +41,7 @@ describe("<Queries />", () => {
       query: GET_QUERIES,
       data: {
         watchedQueries: {
+          __typename: "WatchedQueries",
           queries,
           count: queries.length,
         },
@@ -64,10 +68,13 @@ describe("<Queries />", () => {
   });
 
   test("renders query name", async () => {
+    const user = userEvent.setup();
+
     client.writeQuery({
       query: GET_QUERIES,
       data: {
         watchedQueries: {
+          __typename: "WatchedQueries",
           queries,
           count: queries.length,
         },
@@ -85,7 +92,7 @@ describe("<Queries />", () => {
     expect(within(header).getByText("Unnamed")).toBeInTheDocument();
 
     const sidebar = screen.getByTestId("sidebar");
-    await user.click(within(sidebar).getByText("GetColors"));
+    await act(() => user.click(within(sidebar).getByText("GetColors")));
     await waitFor(() => {
       expect(within(header).getByText("GetColors")).toBeInTheDocument();
     });